+++ scripts/brave_rewards/publisher/common/utils.ts
